FIXED BundleInstaller: move fails if bundle was not installed yet UPDATED leoz-boot version to 0.6-RELEASE

+++ libs/leoz-common/src/main/java/org/deku/leoz/bundle/BundleInstaller.kt
@@ -261,11 +261,14 @@ class BundleInstaller(
                 oldBundlePath.deleteRecursively()
             }
 
-            Files.move(bundlePath.toPath(), oldBundlePath.toPath())
+            if (bundlePath.exists())
+                Files.move(bundlePath.toPath(), oldBundlePath.toPath())
+
             try {
                 Files.move(bundleReadyPath.toPath(), bundlePath.toPath())
             } catch(e: Exception) {
-                Files.move(oldBundlePath.toPath(), bundlePath.toPath())
+                if (oldBundlePath.exists())
+                    Files.move(oldBundlePath.toPath(), bundlePath.toPath())
                 throw e
             }
             oldBundlePath.deleteRecursively()
